+++ motor/__init__.py
@@ -14,7 +14,7 @@
 
 from __future__ import unicode_literals, absolute_import
 
-"""Motor, an asynchronous driver for MongoDB and Tornado."""
+"""Motor, an asynchronous driver for MongoDB."""
 
 import collections
 import functools
@@ -24,7 +24,7 @@ import sys
 import time
 import warnings
 
-from tornado import ioloop, iostream, gen, stack_context, netutil
+from tornado import ioloop, gen, stack_context, netutil
 from tornado.concurrent import Future, TracebackFuture
 import greenlet
 
@@ -36,6 +36,9 @@ except ImportError:
 
 import bson
 import pymongo
+
+from motor.frameworks import tornado as tornado_framework
+from motor.motor_common import callback_type_error
 
 version_tuple = (0, 3)
 
@@ -100,180 +103,9 @@ def check_deprecated_kwargs(kwargs):
             "Motor does not support requests")
 
 
-callback_type_error = TypeError("callback must be a callable")
-
-
-def motor_sock_method(method):
-    """Wrap a MotorSocket method to pause the current greenlet and arrange
-       for the greenlet to be resumed when non-blocking I/O has completed.
-    """
-    @functools.wraps(method)
-    def _motor_sock_method(self, *args, **kwargs):
-        child_gr = greenlet.getcurrent()
-        main = child_gr.parent
-        assert main is not None, "Should be on child greenlet"
-
-        timeout_object = None
-
-        if self.timeout:
-            def timeout_err():
-                # Running on the main greenlet. If a timeout error is thrown,
-                # we raise the exception on the child greenlet. Closing the
-                # IOStream removes callback() from the IOLoop so it isn't
-                # called.
-                self.stream.set_close_callback(None)
-                self.stream.close()
-                child_gr.throw(socket.timeout("timed out"))
-
-            timeout_object = self.stream.io_loop.add_timeout(
-                time.time() + self.timeout, timeout_err)
-
-        # This is run by IOLoop on the main greenlet when operation
-        # completes; switch back to child to continue processing
-        def callback(result=None):
-            self.stream.set_close_callback(None)
-            if timeout_object:
-                self.stream.io_loop.remove_timeout(timeout_object)
-
-            child_gr.switch(result)
-
-        # Run on main greenlet
-        def closed():
-            if timeout_object:
-                self.stream.io_loop.remove_timeout(timeout_object)
-
-            # The child greenlet might have died, e.g.:
-            # - An operation raised an error within PyMongo
-            # - PyMongo closed the MotorSocket in response
-            # - MotorSocket.close() closed the IOStream
-            # - IOStream scheduled this closed() function on the loop
-            # - PyMongo operation completed (with or without error) and
-            #       its greenlet terminated
-            # - IOLoop runs this function
-            if not child_gr.dead:
-                child_gr.throw(socket.error("error"))
-
-        self.stream.set_close_callback(closed)
-
-        try:
-            kwargs['callback'] = callback
-
-            # method is MotorSocket.open(), recv(), etc. method() begins a
-            # non-blocking operation on an IOStream and arranges for
-            # callback() to be executed on the main greenlet once the
-            # operation has completed.
-            method(self, *args, **kwargs)
-
-            # Pause child greenlet until resumed by main greenlet, which
-            # will pass the result of the socket operation (data for recv,
-            # number of bytes written for sendall) to us.
-            return main.switch()
-        except socket.error:
-            raise
-        except IOError as e:
-            # If IOStream raises generic IOError (e.g., if operation
-            # attempted on closed IOStream), then substitute socket.error,
-            # since socket.error is what PyMongo's built to handle. For
-            # example, PyMongo will catch socket.error, close the socket,
-            # and raise AutoReconnect.
-            raise socket.error(str(e))
-
-    return _motor_sock_method
-
-
-class MotorSocket(object):
-    """Replace socket with a class that yields from the current greenlet, if
-    we're on a child greenlet, when making blocking calls, and uses Tornado
-    IOLoop to schedule child greenlet for resumption when I/O is ready.
-
-    We only implement those socket methods actually used by pymongo.
-    """
-    def __init__(
-            self, sock, io_loop, use_ssl,
-            certfile, keyfile, ca_certs, cert_reqs):
-        self.use_ssl = use_ssl
-        self.timeout = None
-        if self.use_ssl:
-            # In Python 3, Tornado's ssl_options_to_context fails if
-            # any options are None.
-            ssl_options = {}
-            if certfile:
-                ssl_options['certfile'] = certfile
-
-            if keyfile:
-                ssl_options['keyfile'] = keyfile
-
-            if ca_certs:
-                ssl_options['ca_certs'] = ca_certs
-
-            if cert_reqs:
-                ssl_options['cert_reqs'] = cert_reqs
-
-            self.stream = iostream.SSLIOStream(
-                sock, ssl_options=ssl_options, io_loop=io_loop)
-        else:
-            self.stream = iostream.IOStream(sock, io_loop=io_loop)
-
-    def setsockopt(self, *args, **kwargs):
-        self.stream.socket.setsockopt(*args, **kwargs)
-
-    def settimeout(self, timeout):
-        # IOStream calls socket.setblocking(False), which does settimeout(0.0).
-        # We must not allow pymongo to set timeout to some other value (a
-        # positive number or None) or the socket will start blocking again.
-        # Instead, we simulate timeouts by interrupting ourselves with
-        # callbacks.
-        self.timeout = timeout
-
-    @motor_sock_method
-    def connect(self, pair, server_hostname=None, callback=None):
-        """
-        :Parameters:
-         - `pair`: A tuple, (host, port)
-        """
-        # 'server_hostname' is used for optional certificate validation.
-        self.stream.connect(pair, callback, server_hostname=server_hostname)
-
-    def sendall(self, data):
-        assert greenlet.getcurrent().parent is not None,\
-            "Should be on child greenlet"
-
-        try:
-            self.stream.write(data)
-        except IOError as e:
-            # PyMongo is built to handle socket.error here, not IOError
-            raise socket.error(str(e))
-
-        if self.stream.closed():
-            # Something went wrong while writing
-            raise socket.error("write error")
-
-    @motor_sock_method
-    def recv(self, num_bytes, callback):
-        self.stream.read_bytes(num_bytes, callback)
-
-    def close(self):
-        sock = self.stream.socket
-        try:
-            self.stream.close()
-        except KeyError:
-            # Tornado's _impl (epoll, kqueue, ...) has already removed this
-            # file descriptor from its dict.
-            pass
-        finally:
-            # Sometimes necessary to avoid ResourceWarnings in Python 3:
-            # specifically, if the fd is closed from the OS's view, then
-            # stream.close() throws an exception, but the socket still has an
-            # fd and so will print a ResourceWarning. In that case, calling
-            # sock.close() directly clears the fd and does not raise an error.
-            if sock:
-                sock.close()
-
-    def fileno(self):
-        return self.stream.socket.fileno()
-
-
 class MotorPool(object):
+    framework = tornado_framework
+
     def __init__(
             self, io_loop, pair, max_size, net_timeout, conn_timeout, use_ssl,
             use_greenlets,
@@ -281,7 +113,7 @@ class MotorPool(object):
             ssl_cert_reqs=None, ssl_ca_certs=None,
             wait_queue_timeout=None, wait_queue_multiple=None):
         """
-        A pool of MotorSockets.
+        A connection pool that uses Motor's framework-specific sockets.
 
         :Parameters:
           - `io_loop`: An IOLoop instance
@@ -419,7 +251,7 @@ class MotorPool(object):
             sock = None
             try:
                 sock = socket.socket(af, socktype, proto)
-                motor_sock = MotorSocket(
+                motor_sock = self.framework.create_socket(
                     sock, self.io_loop, use_ssl=self.use_ssl,
                     certfile=self.ssl_certfile, keyfile=self.ssl_keyfile,
                     ca_certs=self.ssl_ca_certs, cert_reqs=self.ssl_cert_reqs)
@@ -608,33 +440,6 @@ class MotorPool(object):
                 self.max_size, self.wait_queue_timeout))
 
 
-def motor_coroutine(f):
-    """A coroutine that accepts an optional callback.
-
-    Given a callback, the function returns None, and the callback is run
-    with (result, error). Without a callback the function returns a Future.
-    """
-    coro = gen.coroutine(f)
-
-    @functools.wraps(f)
-    def wrapper(*args, **kwargs):
-        callback = kwargs.pop('callback', None)
-        if callback and not callable(callback):
-            raise callback_type_error
-        future = coro(*args, **kwargs)
-        if callback:
-            def _callback(future):
-                try:
-                    result = future.result()
-                    callback(result, None)
-                except Exception as e:
-                    callback(None, e)
-            future.add_done_callback(_callback)
-        else:
-            return future
-    return wrapper
-
-
 def mangle_delegate_name(motor_class, name):
     if name.startswith('__') and not name.endswith("__"):
         # Mangle, e.g. Cursor.__die -> Cursor._Cursor__die
@@ -784,7 +589,7 @@ class WrapAsync(WrapBase):
         original_class = self.original_class
 
         @functools.wraps(async_method)
-        @motor_coroutine
+        @cls.framework.coroutine
         def wrapper(self, *args, **kwargs):
             result = yield async_method(self, *args, **kwargs)
 
@@ -920,6 +725,7 @@ class MotorMeta(type):
     """
     def __new__(cls, class_name, bases, attrs):
         new_class = type.__new__(cls, class_name, bases, attrs)
+        new_class.framework = tornado_framework
 
         # If new_class has no __delegate_class__, then it's a base like
         # MotorClientBase; don't try to update its attrs, we'll use them
@@ -1001,7 +807,7 @@ class MotorClientBase(MotorBase):
 
     __getitem__ = __getattr__
 
-    @motor_coroutine
+    @tornado_framework.coroutine
     def copy_database(
             self, from_name, to_name, from_host=None, username=None,
             password=None):
@@ -1130,7 +936,7 @@ class MotorClient(MotorClientBase):
         kwargs['_event_class'] = event_class
         super(MotorClient, self).__init__(io_loop, *args, **kwargs)
 
-    @motor_coroutine
+    @tornado_framework.coroutine
     def open(self):
         """Connect to the server.
 
@@ -1205,7 +1011,7 @@ class MotorReplicaSetClient(MotorClientBase):
 
         super(MotorReplicaSetClient, self).__init__(io_loop, *args, **kwargs)
 
-    @motor_coroutine
+    @tornado_framework.coroutine
     def open(self):
         """Connect to the server.
 
@@ -1509,7 +1315,7 @@ class MotorCollection(MotorBase):
         cursor = self.delegate.find(*args, **kwargs)
         return MotorCursor(cursor, self)
 
-    @motor_coroutine
+    @tornado_framework.coroutine
     def parallel_scan(self, num_cursors, **kwargs):
         """Scan this entire collection in parallel.
 
@@ -1808,7 +1614,7 @@ class _MotorBaseCursor(MotorBase):
             # Complete
             add_callback(functools.partial(callback, None, None))
 
-    @motor_coroutine
+    @tornado_framework.coroutine
     def to_list(self, length):
         """Get a list of documents.
 
@@ -1884,7 +1690,7 @@ class _MotorBaseCursor(MotorBase):
     def get_io_loop(self):
         return self.collection.get_io_loop()
 
-    @motor_coroutine
+    @tornado_framework.coroutine
     def close(self):
         """Explicitly kill this cursor on the server. If iterating with
         :meth:`each`, cease.
@@ -2284,7 +2090,7 @@ class MotorGridOut(object):
 
         return getattr(self.delegate, item)
 
-    @motor_coroutine
+    @tornado_framework.coroutine
     def open(self):
         """Retrieve this file's attributes from the server.
 
@@ -2303,7 +2109,7 @@ class MotorGridOut(object):
     def get_io_loop(self):
         return self.io_loop
 
-    @motor_coroutine
+    @tornado_framework.coroutine
     def stream_to_handler(self, request_handler):
         """Write the contents of this file to a
         :class:`tornado.web.RequestHandler`. This method calls `flush` on
@@ -2480,7 +2286,7 @@ class MotorGridFS(object):
     def get_io_loop(self):
         return self.io_loop
 
-    @motor_coroutine
+    @tornado_framework.coroutine
     def put(self, data, **kwargs):
         """Put data into GridFS as a new file.
